server/requirements.txt
openenv-core>=0.2.3
fastapi>=0.104.0
uvicorn>=0.24.0
transformer-lens>=2.0.0
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
datasets>=2.18.0
wandb>=0.16.0
omegaconf>=2.3.0
rich>=13.7.0
numpy>=1.26.0
tqdm>=4.66.0
